.travis-requirements.txt
coverage>=3.7.1
coveralls>=0.4.2
django-flat-theme>=1.1.1
pytz>=2015.6